Add sample wasm-bindgen configuration
Wrap `tlock_age` methods with wasm-bindgen to allow for a use in JavaScript and TypeScript.
This is done through an optional feature `js`.
Armoring is supported.

Methods use a Uint8Array containing the full source to encrypt or decrypt. Result is returned as Uint8Array.

Two targets have been tested
- NodeJS `wasm-pack build --target nodejs -- --features js`. Then `const tlock_age = require('./pkg')`
- Web `wasm-pack build --target web -- --features js`. You can then point your bundler to use `./pkg`. Don't forget to call the default export `init()` to load the module in browser.